Pull out common interface into separate file
