+++ log/log.go
@@ -1,0 +1,24 @@
+package log
+
+/* CHECKLIST
+ * [x] Uses interfaces as appropriate
+ * [x] Private package variables use underscore prefix
+ * [x] All parameters validated
+ * [x] All errors handled
+ * [x] Reviewed for concurrency safety
+ * [x] Code complete
+ * [x] Full test coverage
+ */
+
+type Fields map[string]interface{}
+
+type Logger interface {
+	Debug(message string)
+	Info(message string)
+	Warn(message string)
+	Error(message string)
+
+	WithError(err error) Logger
+	WithField(key string, value interface{}) Logger
+	WithFields(fields Fields) Logger
+}
