+++ log/standard.go
@@ -22,19 +22,6 @@ import (
 	"github.com/tidepool-org/platform/version"
 )
 
-type Fields map[string]interface{}
-
-type Logger interface {
-	Debug(message string)
-	Info(message string)
-	Warn(message string)
-	Error(message string)
-
-	WithError(err error) Logger
-	WithField(key string, value interface{}) Logger
-	WithFields(fields Fields) Logger
-}
-
 type Standard struct {
 	logger              *logrus.Logger
 	fields              logrus.Fields
